+++ test/acceptance/codecept.Playwright.js
@@ -11,6 +11,10 @@ module.exports.config = {
       show: false,
       restart: process.env.BROWSER_RESTART || false,
       browser: process.env.BROWSER || 'chromium',
+      ignoreHTTPSErrors: true,
+      webkit: {
+        ignoreHTTPSErrors: true,
+      },
     },
     JSONResponse: {
       requestHelper: 'Playwright',