+++ test/data/sandbox/configs/bootstrap/invalid_require_test.js
@@ -5,5 +5,5 @@ Feature('Filesystem');
 Scenario('check current dir', ({ I }) => {
   I.amInPath('.');
   I.say('hello world');
-  I.seeFile('codecept.json');
+  I.seeFile('codecept.js');
 });