+++ test/data/sandbox/fs_test.js
@@ -3,5 +3,5 @@ Feature('Filesystem').tag('main');
 Scenario('check current dir', ({ I }) => {
   I.amInPath('.');
   I.say('hello world');
-  I.seeFile('codecept.json');
+  I.seeFile('codecept.js');
 }).tag('slow').tag('@important');