+++ test/helper/JSONResponse_test.js
@@ -83,11 +83,42 @@ describe('JSONResponse', () => {
           { id: 1, author: 'davert' },
         ],
       });
-      expect(() => I.seeResponseContainsJson({ posts: [{ id: 2, author: 'boss' }] })).to.throw('AssertionError');
+      expect(() => I.seeResponseContainsJson({ posts: [{ id: 2, author: 'boss' }] })).to.throw('expected { …(2) } to deeply match { Object (posts) }');
+    });
+
+    it('should check for json inclusion - returned Array', () => {
+      const arrayData = [{ ...data }];
+      restHelper.config.onResponse({ data: arrayData });
+      I.seeResponseContainsJson({
+        posts: [
+          { id: 2 },
+        ],
+      });
+      I.seeResponseContainsJson({
+        posts: [
+          { id: 1, author: 'davert' },
+        ],
+      });
+      expect(() => I.seeResponseContainsJson({ posts: [{ id: 2, author: 'boss' }] })).to.throw('No elements in array matched {"posts":[{"id":2,"author":"boss"}]}');
+    });
+
+    it('should check for json inclusion - returned Array of 2 items', () => {
+      const arrayData = [{ ...data }, { posts: { id: 3 } }];
+      restHelper.config.onResponse({ data: arrayData });
+      I.seeResponseContainsJson({
+        posts: { id: 3 },
+      });
     });
 
     it('should simply check for json inclusion', () => {
       restHelper.config.onResponse({ data: { user: { name: 'jon', email: 'jon@doe.com' } } });
+      I.seeResponseContainsJson({ user: { name: 'jon' } });
+      I.dontSeeResponseContainsJson({ user: { name: 'jo' } });
+      I.dontSeeResponseContainsJson({ name: 'joe' });
+    });
+
+    it('should simply check for json inclusion - returned Array', () => {
+      restHelper.config.onResponse({ data: [{ user: { name: 'jon', email: 'jon@doe.com' } }] });
       I.seeResponseContainsJson({ user: { name: 'jon' } });
       I.dontSeeResponseContainsJson({ user: { name: 'jo' } });
       I.dontSeeResponseContainsJson({ name: 'joe' });
@@ -98,8 +129,18 @@ describe('JSONResponse', () => {
       I.seeResponseEquals({ user: 1 });
     });
 
+    it('should simply check for json equality - returned Array', () => {
+      restHelper.config.onResponse({ data: [{ user: 1 }] });
+      I.seeResponseEquals([{ user: 1 }]);
+    });
+
     it('should check json contains keys', () => {
       restHelper.config.onResponse({ data: { user: 1, post: 2 } });
+      I.seeResponseContainsKeys(['user', 'post']);
+    });
+
+    it('should check json contains keys - returned Array', () => {
+      restHelper.config.onResponse({ data: [{ user: 1, post: 2 }] });
       I.seeResponseContainsKeys(['user', 'post']);
     });
 
@@ -124,7 +165,7 @@ describe('JSONResponse', () => {
           name: joi.string(),
         }),
       });
-      const fn = (joi) => {
+      const fn = () => {
         return schema;
       };
       I.seeResponseMatchesJsonSchema(fn);