+++ typings/tests/helpers/WebDriverIO.types.ts
@@ -4,10 +4,10 @@ const str_wd = 'text';
 const num_wd = 1;
 
 wd.amOnPage(); // $ExpectError
-wd.amOnPage(''); // $ExpectType Promise<any>
+wd.amOnPage(''); // $ExpectType void
 
 wd.click(); // $ExpectError
-wd.click('div'); // $ExpectType Promise<any>
+wd.click('div'); // $ExpectType void
 wd.click({ css: 'div' });
 wd.click({ xpath: '//div' });
 wd.click({ name: 'div' });
@@ -25,7 +25,7 @@ wd.click('div', { android: '//div' });
 wd.click('div', { ios: '//div' });
 
 wd.forceClick(); // $ExpectError
-wd.forceClick('div'); // $ExpectType Promise<any>
+wd.forceClick('div'); // $ExpectType void
 wd.forceClick({ css: 'div' });
 wd.forceClick({ xpath: '//div' });
 wd.forceClick({ name: 'div' });
@@ -43,7 +43,7 @@ wd.forceClick('div', { android: '//div' });
 wd.forceClick('div', { ios: '//div' });
 
 wd.doubleClick(); // $ExpectError
-wd.doubleClick('div'); // $ExpectType Promise<any>
+wd.doubleClick('div'); // $ExpectType void
 wd.doubleClick({ css: 'div' });
 wd.doubleClick({ xpath: '//div' });
 wd.doubleClick({ name: 'div' });
@@ -61,7 +61,7 @@ wd.doubleClick('div', { android: '//div' });
 wd.doubleClick('div', { ios: '//div' });
 
 wd.rightClick(); // $ExpectError
-wd.rightClick('div'); // $ExpectType Promise<any>
+wd.rightClick('div'); // $ExpectType void
 wd.rightClick({ css: 'div' });
 wd.rightClick({ xpath: '//div' });
 wd.rightClick({ name: 'div' });
@@ -80,7 +80,7 @@ wd.rightClick('div', { ios: '//div' });
 
 wd.fillField(); // $ExpectError
 wd.fillField('div'); // $ExpectError
-wd.fillField('div', str_wd); // $ExpectType Promise<any>
+wd.fillField('div', str_wd); // $ExpectType void
 wd.fillField({ css: 'div' }, str_wd);
 wd.fillField({ xpath: '//div' }, str_wd);
 wd.fillField({ name: 'div' }, str_wd);
@@ -91,7 +91,7 @@ wd.fillField(locate('div'), str_wd);
 
 wd.appendField(); // $ExpectError
 wd.appendField('div'); // $ExpectError
-wd.appendField('div', str_wd); // $ExpectType Promise<any>
+wd.appendField('div', str_wd); // $ExpectType void
 wd.appendField({ css: 'div' }, str_wd);
 wd.appendField({ xpath: '//div' }, str_wd);
 wd.appendField({ name: 'div' }, str_wd);
@@ -111,134 +111,134 @@ wd.clearField({ ios: 'div' });
 
 wd.selectOption(); // $ExpectError
 wd.selectOption('div'); // $ExpectError
-wd.selectOption('div', str_wd); // $ExpectType Promise<any>
+wd.selectOption('div', str_wd); // $ExpectType void
 
 wd.attachFile(); // $ExpectError
 wd.attachFile('div'); // $ExpectError
-wd.attachFile('div', str_wd); // $ExpectType Promise<any>
+wd.attachFile('div', str_wd); // $ExpectType void
 
 wd.checkOption(); // $ExpectError
-wd.checkOption('div'); // $ExpectType Promise<any>
+wd.checkOption('div'); // $ExpectType void
 
 wd.uncheckOption(); // $ExpectError
-wd.uncheckOption('div'); // $ExpectType Promise<any>
+wd.uncheckOption('div'); // $ExpectType void
 
 wd.seeInTitle(); // $ExpectError
-wd.seeInTitle(str_wd); // $ExpectType Promise<any>
+wd.seeInTitle(str_wd); // $ExpectType void
 
 wd.seeTitleEquals(); // $ExpectError
-wd.seeTitleEquals(str_wd); // $ExpectType Promise<any>
+wd.seeTitleEquals(str_wd); // $ExpectType void
 
 wd.dontSeeInTitle(); // $ExpectError
-wd.dontSeeInTitle(str_wd); // $ExpectType Promise<any>
+wd.dontSeeInTitle(str_wd); // $ExpectType void
 
 wd.see(); // $ExpectError
-wd.see(str_wd); // $ExpectType Promise<any>
-wd.see(str_wd, 'div'); // $ExpectType Promise<any>
+wd.see(str_wd); // $ExpectType void
+wd.see(str_wd, 'div'); // $ExpectType void
 
 wd.dontSee(); // $ExpectError
-wd.dontSee(str_wd); // $ExpectType Promise<any>
-wd.dontSee(str_wd, 'div'); // $ExpectType Promise<any>
+wd.dontSee(str_wd); // $ExpectType void
+wd.dontSee(str_wd, 'div'); // $ExpectType void
 
 wd.seeTextEquals(); // $ExpectError
-wd.seeTextEquals(str_wd); // $ExpectType Promise<any>
-wd.seeTextEquals(str_wd, 'div'); // $ExpectType Promise<any>
+wd.seeTextEquals(str_wd); // $ExpectType void
+wd.seeTextEquals(str_wd, 'div'); // $ExpectType void
 
 wd.seeInField(); // $ExpectError
 wd.seeInField('div'); // $ExpectError
-wd.seeInField('div', str_wd); // $ExpectType Promise<any>
+wd.seeInField('div', str_wd); // $ExpectType void
 
 wd.dontSeeInField(); // $ExpectError
 wd.dontSeeInField('div'); // $ExpectError
-wd.dontSeeInField('div', str_wd); // $ExpectType Promise<any>
+wd.dontSeeInField('div', str_wd); // $ExpectType void
 
 wd.seeCheckboxIsChecked(); // $ExpectError
-wd.seeCheckboxIsChecked('div'); // $ExpectType Promise<any>
+wd.seeCheckboxIsChecked('div'); // $ExpectType void
 
 wd.dontSeeCheckboxIsChecked(); // $ExpectError
-wd.dontSeeCheckboxIsChecked('div'); // $ExpectType Promise<any>
+wd.dontSeeCheckboxIsChecked('div'); // $ExpectType void
 
 wd.seeElement(); // $ExpectError
-wd.seeElement('div'); // $ExpectType Promise<any>
+wd.seeElement('div'); // $ExpectType void
 
 wd.dontSeeElement(); // $ExpectError
-wd.dontSeeElement('div'); // $ExpectType Promise<any>
+wd.dontSeeElement('div'); // $ExpectType void
 
 wd.seeElementInDOM(); // $ExpectError
-wd.seeElementInDOM('div'); // $ExpectType Promise<any>
+wd.seeElementInDOM('div'); // $ExpectType void
 
 wd.dontSeeElementInDOM(); // $ExpectError
-wd.dontSeeElementInDOM('div'); // $ExpectType Promise<any>
+wd.dontSeeElementInDOM('div'); // $ExpectType void
 
 wd.seeInSource(); // $ExpectError
-wd.seeInSource(str_wd); // $ExpectType Promise<any>
+wd.seeInSource(str_wd); // $ExpectType void
 
 wd.dontSeeInSource(); // $ExpectError
-wd.dontSeeInSource(str_wd); // $ExpectType Promise<any>
+wd.dontSeeInSource(str_wd); // $ExpectType void
 
 wd.seeNumberOfElements(); // $ExpectError
 wd.seeNumberOfElements('div'); // $ExpectError
-wd.seeNumberOfElements('div', num_wd); // $ExpectType Promise<any>
+wd.seeNumberOfElements('div', num_wd); // $ExpectType void
 
 wd.seeNumberOfVisibleElements(); // $ExpectError
 wd.seeNumberOfVisibleElements('div'); // $ExpectError
-wd.seeNumberOfVisibleElements('div', num_wd); // $ExpectType Promise<any>
+wd.seeNumberOfVisibleElements('div', num_wd); // $ExpectType void
 
 wd.seeCssPropertiesOnElements(); // $ExpectError
 wd.seeCssPropertiesOnElements('div'); // $ExpectError
-wd.seeCssPropertiesOnElements('div', str_wd); // $ExpectType Promise<any>
+wd.seeCssPropertiesOnElements('div', str_wd); // $ExpectType void
 
 wd.seeAttributesOnElements(); // $ExpectError
 wd.seeAttributesOnElements('div'); // $ExpectError
-wd.seeAttributesOnElements('div', str_wd); // $ExpectType Promise<any>
+wd.seeAttributesOnElements('div', str_wd); // $ExpectType void
 
 wd.seeInCurrentUrl(); // $ExpectError
-wd.seeInCurrentUrl(str_wd); // $ExpectType Promise<any>
+wd.seeInCurrentUrl(str_wd); // $ExpectType void
 
 wd.seeCurrentUrlEquals(); // $ExpectError
-wd.seeCurrentUrlEquals(str_wd); // $ExpectType Promise<any>
+wd.seeCurrentUrlEquals(str_wd); // $ExpectType void
 
 wd.dontSeeInCurrentUrl(); // $ExpectError
-wd.dontSeeInCurrentUrl(str_wd); // $ExpectType Promise<any>
+wd.dontSeeInCurrentUrl(str_wd); // $ExpectType void
 
 wd.dontSeeCurrentUrlEquals(); // $ExpectError
-wd.dontSeeCurrentUrlEquals(str_wd); // $ExpectType Promise<any>
+wd.dontSeeCurrentUrlEquals(str_wd); // $ExpectType void
 
 wd.executeScript(); // $ExpectError
-wd.executeScript(str_wd); // $ExpectType Promise<any>
-wd.executeScript(() => {}); // $ExpectType Promise<any>
-wd.executeScript(() => {}, {}); // $ExpectType Promise<any>
+wd.executeScript(str_wd); // $ExpectType void
+wd.executeScript(() => {}); // $ExpectType void
+wd.executeScript(() => {}, {}); // $ExpectType void
 
 wd.executeAsyncScript(); // $ExpectError
-wd.executeAsyncScript(str_wd); // $ExpectType Promise<any>
-wd.executeAsyncScript(() => {}); // $ExpectType Promise<any>
-wd.executeAsyncScript(() => {}, {}); // $ExpectType Promise<any>
+wd.executeAsyncScript(str_wd); // $ExpectType void
+wd.executeAsyncScript(() => {}); // $ExpectType void
+wd.executeAsyncScript(() => {}, {}); // $ExpectType void
 
 wd.scrollIntoView(); // $ExpectError
 wd.scrollIntoView('div'); // $ExpectError
 wd.scrollIntoView('div', { behavior: 'auto', block: 'center', inline: 'center' });
 
 wd.scrollTo(); // $ExpectError
-wd.scrollTo('div'); // $ExpectType Promise<any>
-wd.scrollTo('div', num_wd, num_wd); // $ExpectType Promise<any>
+wd.scrollTo('div'); // $ExpectType void
+wd.scrollTo('div', num_wd, num_wd); // $ExpectType void
 
 wd.moveCursorTo(); // $ExpectError
-wd.moveCursorTo('div'); // $ExpectType Promise<any>
-wd.moveCursorTo('div', num_wd, num_wd); // $ExpectType Promise<any>
+wd.moveCursorTo('div'); // $ExpectType void
+wd.moveCursorTo('div', num_wd, num_wd); // $ExpectType void
 
 wd.saveScreenshot(); // $ExpectError
-wd.saveScreenshot(str_wd); // $ExpectType Promise<any>
-wd.saveScreenshot(str_wd, true); // $ExpectType Promise<any>
+wd.saveScreenshot(str_wd); // $ExpectType void
+wd.saveScreenshot(str_wd, true); // $ExpectType void
 
 wd.setCookie(); // $ExpectError
-wd.setCookie({ name: str_wd, value: str_wd }); // $ExpectType Promise<any>
-wd.setCookie([{ name: str_wd, value: str_wd }]); // $ExpectType Promise<any>
+wd.setCookie({ name: str_wd, value: str_wd }); // $ExpectType void
+wd.setCookie([{ name: str_wd, value: str_wd }]); // $ExpectType void
 
-wd.clearCookie(); // $ExpectType Promise<any>
-wd.clearCookie(str_wd); // $ExpectType Promise<any>
+wd.clearCookie(); // $ExpectType void
+wd.clearCookie(str_wd); // $ExpectType void
 
 wd.seeCookie(); // $ExpectError
-wd.seeCookie(str_wd); // $ExpectType Promise<any>
+wd.seeCookie(str_wd); // $ExpectType void
 
 wd.acceptPopup(); // $ExpectType void
 
@@ -248,116 +248,116 @@ wd.seeInPopup(); // $ExpectError
 wd.seeInPopup(str_wd); // $ExpectType void
 
 wd.pressKeyDown(); // $ExpectError
-wd.pressKeyDown(str_wd); // $ExpectType Promise<any>
+wd.pressKeyDown(str_wd); // $ExpectType void
 
 wd.pressKeyUp(); // $ExpectError
-wd.pressKeyUp(str_wd); // $ExpectType Promise<any>
+wd.pressKeyUp(str_wd); // $ExpectType void
 
 wd.pressKey(); // $ExpectError
-wd.pressKey(str_wd); // $ExpectType Promise<any>
+wd.pressKey(str_wd); // $ExpectType void
 
 wd.type(); // $ExpectError
-wd.type(str_wd); // $ExpectType Promise<any>
+wd.type(str_wd); // $ExpectType void
 
 wd.resizeWindow(); // $ExpectError
 wd.resizeWindow(num_wd); // $ExpectError
-wd.resizeWindow(num_wd, num_wd); // $ExpectType Promise<any>
+wd.resizeWindow(num_wd, num_wd); // $ExpectType void
 
 wd.dragAndDrop(); // $ExpectError
 wd.dragAndDrop('div'); // $ExpectError
-wd.dragAndDrop('div', 'div'); // $ExpectType Promise<any>
+wd.dragAndDrop('div', 'div'); // $ExpectType void
 
 wd.dragSlider(); // $ExpectError
-wd.dragSlider('div', num_wd); // $ExpectType Promise<any>
+wd.dragSlider('div', num_wd); // $ExpectType void
 
 wd.switchToWindow(); // $ExpectError
 wd.switchToWindow(str_wd); // $ExpectType void
 
-wd.closeOtherTabs(); // $ExpectType Promise<any>
+wd.closeOtherTabs(); // $ExpectType void
 
 wd.wait(); // $ExpectError
-wd.wait(num_wd); // $ExpectType Promise<any>
+wd.wait(num_wd); // $ExpectType void
 
 wd.waitForEnabled(); // $ExpectError
-wd.waitForEnabled('div'); // $ExpectType Promise<any>
-wd.waitForEnabled('div', num_wd); // $ExpectType Promise<any>
+wd.waitForEnabled('div'); // $ExpectType void
+wd.waitForEnabled('div', num_wd); // $ExpectType void
 
 wd.waitForElement(); // $ExpectError
-wd.waitForElement('div'); // $ExpectType Promise<any>
-wd.waitForElement('div', num_wd); // $ExpectType Promise<any>
+wd.waitForElement('div'); // $ExpectType void
+wd.waitForElement('div', num_wd); // $ExpectType void
 
 wd.waitForClickable(); // $ExpectError
-wd.waitForClickable('div'); // $ExpectType Promise<any>
-wd.waitForClickable('div', num_wd); // $ExpectType Promise<any>
+wd.waitForClickable('div'); // $ExpectType void
+wd.waitForClickable('div', num_wd); // $ExpectType void
 
 wd.waitForVisible(); // $ExpectError
-wd.waitForVisible('div'); // $ExpectType Promise<any>
-wd.waitForVisible('div', num_wd); // $ExpectType Promise<any>
+wd.waitForVisible('div'); // $ExpectType void
+wd.waitForVisible('div', num_wd); // $ExpectType void
 
 wd.waitForInvisible(); // $ExpectError
-wd.waitForInvisible('div'); // $ExpectType Promise<any>
-wd.waitForInvisible('div', num_wd); // $ExpectType Promise<any>
+wd.waitForInvisible('div'); // $ExpectType void
+wd.waitForInvisible('div', num_wd); // $ExpectType void
 
 wd.waitToHide(); // $ExpectError
-wd.waitToHide('div'); // $ExpectType Promise<any>
-wd.waitToHide('div', num_wd); // $ExpectType Promise<any>
+wd.waitToHide('div'); // $ExpectType void
+wd.waitToHide('div', num_wd); // $ExpectType void
 
 wd.waitForDetached(); // $ExpectError
-wd.waitForDetached('div'); // $ExpectType Promise<any>
-wd.waitForDetached('div', num_wd); // $ExpectType Promise<any>
+wd.waitForDetached('div'); // $ExpectType void
+wd.waitForDetached('div', num_wd); // $ExpectType void
 
 wd.waitForFunction(); // $ExpectError
-wd.waitForFunction('div'); // $ExpectType Promise<any>
-wd.waitForFunction(() => {}); // $ExpectType Promise<any>
-wd.waitForFunction(() => {}, [num_wd], num_wd); // $ExpectType Promise<any>
-wd.waitForFunction(() => {}, [str_wd], num_wd); // $ExpectType Promise<any>
+wd.waitForFunction('div'); // $ExpectType void
+wd.waitForFunction(() => {}); // $ExpectType void
+wd.waitForFunction(() => {}, [num_wd], num_wd); // $ExpectType void
+wd.waitForFunction(() => {}, [str_wd], num_wd); // $ExpectType void
 
 wd.waitInUrl(); // $ExpectError
-wd.waitInUrl(str_wd); // $ExpectType Promise<any>
-wd.waitInUrl(str_wd, num_wd); // $ExpectType Promise<any>
+wd.waitInUrl(str_wd); // $ExpectType void
+wd.waitInUrl(str_wd, num_wd); // $ExpectType void
 
 wd.waitForText(); // $ExpectError
-wd.waitForText(str_wd); // $ExpectType Promise<any>
-wd.waitForText(str_wd, num_wd, str_wd); // $ExpectType Promise<any>
+wd.waitForText(str_wd); // $ExpectType void
+wd.waitForText(str_wd, num_wd, str_wd); // $ExpectType void
 
 wd.waitForValue(); // $ExpectError
 wd.waitForValue(str_wd); // $ExpectError
-wd.waitForValue(str_wd, str_wd); // $ExpectType Promise<any>
-wd.waitForValue(str_wd, str_wd, num_wd); // $ExpectType Promise<any>
+wd.waitForValue(str_wd, str_wd); // $ExpectType void
+wd.waitForValue(str_wd, str_wd, num_wd); // $ExpectType void
 
 wd.waitNumberOfVisibleElements(); // $ExpectError
 wd.waitNumberOfVisibleElements('div'); // $ExpectError
-wd.waitNumberOfVisibleElements(str_wd, num_wd); // $ExpectType Promise<any>
-wd.waitNumberOfVisibleElements(str_wd, num_wd, num_wd); // $ExpectType Promise<any>
+wd.waitNumberOfVisibleElements(str_wd, num_wd); // $ExpectType void
+wd.waitNumberOfVisibleElements(str_wd, num_wd, num_wd); // $ExpectType void
 
 wd.waitUrlEquals(); // $ExpectError
-wd.waitUrlEquals(str_wd); // $ExpectType Promise<any>
-wd.waitUrlEquals(str_wd, num_wd); // $ExpectType Promise<any>
+wd.waitUrlEquals(str_wd); // $ExpectType void
+wd.waitUrlEquals(str_wd, num_wd); // $ExpectType void
 
-wd.switchTo(); // $ExpectType Promise<any>
-wd.switchTo('div'); // $ExpectType Promise<any>
+wd.switchTo(); // $ExpectType void
+wd.switchTo('div'); // $ExpectType void
 
-wd.switchToNextTab(num_wd, num_wd); // $ExpectType Promise<any>
+wd.switchToNextTab(num_wd, num_wd); // $ExpectType void
 
-wd.switchToPreviousTab(num_wd, num_wd); // $ExpectType Promise<any>
+wd.switchToPreviousTab(num_wd, num_wd); // $ExpectType void
 
-wd.closeCurrentTab(); // $ExpectType Promise<any>
+wd.closeCurrentTab(); // $ExpectType void
 
-wd.openNewTab(); // $ExpectType Promise<any>
+wd.openNewTab(); // $ExpectType void
 
-wd.refreshPage(); // $ExpectType Promise<any>
+wd.refreshPage(); // $ExpectType void
 
-wd.scrollPageToTop(); // $ExpectType Promise<any>
+wd.scrollPageToTop(); // $ExpectType void
 
-wd.scrollPageToBottom(); // $ExpectType Promise<any>
+wd.scrollPageToBottom(); // $ExpectType void
 
 wd.setGeoLocation(); // $ExpectError
 wd.setGeoLocation(num_wd); // $ExpectError
-wd.setGeoLocation(num_wd, num_wd); // $ExpectType Promise<any>
-wd.setGeoLocation(num_wd, num_wd, num_wd); // $ExpectType Promise<any>
+wd.setGeoLocation(num_wd, num_wd); // $ExpectType void
+wd.setGeoLocation(num_wd, num_wd, num_wd); // $ExpectType void
 
 wd.dontSeeCookie(); // $ExpectError
-wd.dontSeeCookie(str_wd); // $ExpectType Promise<any>
+wd.dontSeeCookie(str_wd); // $ExpectType void
 
 wd.dragAndDrop(); // $ExpectError
 wd.dragAndDrop('#dragHandle'); // $ExpectError